pipeline/requirements.txt
scenedetect[opencv]
google-generativeai
anthropic
httpx
python-dotenv
ffmpeg-python
psycopg2-binary
